Feature/rubocop (#56)
* added generalized file

* fine tuned a few things on rubocop config

* simplified the rubocop config

* from 330 warnings to 213

* disabled: Assignment Branch Condition size for call is too high

213 => 207

* now with 195

* 175

* Fixed an issue where device could not be fetch and application would crash
Reduce the error count to 102 on windows

* Fixed spec of github

* Added missing question mark on method

* replaced ors and ands

* 88

* 83 on windows

* 69 on windows

* 43 on windows

* fixed typo

* 38

* Removed @@ variables

* 16 on windows

* disabled perceived complexity, 11 on osX

* freezed constant

* freezed constant

* 7

* update readme

* 4

* Added a check for the existance of a settings gradle file. fixes #54

* Fixed the utils bug

+++ lib/dryrun/android_project.rb
@@ -1,4 +1,4 @@
-require 'oga'
+﻿require 'oga'
 require 'fileutils'
 require 'tempfile'
 require 'find'
@@ -6,12 +6,12 @@ require_relative 'dryrun_utils'
 
 module Dryrun
   class AndroidProject
-    def initialize(path, custom_app_path, custom_module, flavour)
-
+    def initialize(path, custom_app_path, custom_module, flavour, device)
       @custom_app_path = custom_app_path
       @custom_module = custom_module
-      @base_path = @custom_app_path? File.join(path, @custom_app_path) : path
+      @base_path = @custom_app_path ? File.join(path, @custom_app_path) : path
       @flavour = flavour
+      @device = device
 
       @settings_gradle_path = settings_gradle_file
       @main_gradle_file = main_gradle_file
@@ -27,7 +27,7 @@ module Dryrun
       full_custom_path = @base_path
       settings_path = settings_gradle_file(full_custom_path)
       main_gradle_path = main_gradle_file(full_custom_path)
-      return unless is_valid(main_gradle_path)
+      return unless valid?(main_gradle_path)
 
       @settings_gradle_path = settings_path
       @main_gradle_file = main_gradle_file
@@ -40,20 +40,19 @@ module Dryrun
       file_name = 'local.properties'
 
       File.delete(file_name) if File.exist?(file_name)
-      if !Gem.win_platform?
-        DryrunUtils.execute("touch #{file_name}")
-      end
+
+      DryrunUtils.execute("touch #{file_name}") unless Gem.win_platform?
     end
 
     def remove_application_id
       # Open temporary file
-      tmp = Tempfile.new("extract")
+      tmp = Tempfile.new('extract')
 
       file = "#{@path_to_sample}/build.gradle"
 
       # Write good lines to temporary file
-      File.open(file, 'r') do |file|
-        file.each do |l| tmp << l unless l.include? 'applicationId'
+      File.open(file, 'r') do |f|
+        f.each do |l| tmp << l unless l.include? 'applicationId'
         end
       end
       tmp.close
@@ -70,16 +69,17 @@ module Dryrun
       File.join(path, 'build.gradle')
     end
 
-    def is_valid(main_gradle_file = @main_gradle_file)
+    def valid?(main_gradle_file = @main_gradle_file)
       File.exist?(main_gradle_file)
     end
 
     def find_modules
       return [] unless is_valid
+      return [] unless File.exist?(@settings_gradle_path)
 
-      content = File.open(@settings_gradle_path, "rb").read
+      content = File.open(@settings_gradle_path, 'rb').read
       modules = content.scan(/'([^']*)'/)
-      modules.each {|replacement| replacement.first.gsub!(':', '/')}
+      modules.each { |replacement| replacement.first.tr!(':', '/') }
     end
 
     def install
@@ -87,24 +87,24 @@ module Dryrun
 
       path, execute_line = sample_project
 
-      if path == false and execute_line == false
+      if path == false && execute_line == false
         puts "Couldn't open the sample project, sorry!".red
         exit 1
       end
 
-      builder = "gradle"
+      builder = 'gradle'
 
       if File.exist?('gradlew')
         if !Gem.win_platform?
           DryrunUtils.execute('chmod +x gradlew')
         else
-          DryrunUtils.execute("icacls gradlew /T")
+          DryrunUtils.execute('icacls gradlew /T')
         end
         builder = './gradlew'
       end
 
       # Generate the gradle/ folder
-      DryrunUtils.execute('gradle wrap') if File.exist?('gradlew') and !is_gradle_wrapped
+      DryrunUtils.execute('gradle wrap') if File.exist?('gradlew') && !gradle_wrapped?
 
       remove_application_id
       remove_local_properties
@@ -115,17 +115,17 @@ module Dryrun
       else
         DryrunUtils.execute("#{builder} clean")
 
-        if Dryrun::MainApp.getDevice != nil
-          puts "#{builder} install#{@flavour}Debug"
-          DryrunUtils.execute("#{builder} install#{@flavour}Debug")
-        else
-          puts 'No devices picked, proceeding to assemble instead'
+        if @device.nil?
+          puts 'No devices picked/available, proceeding with assemble instead'.green
           puts "#{builder} assemble#{@flavour}Debug"
           DryrunUtils.execute("#{builder} assemble#{@flavour}Debug")
+        else
+          puts "#{builder} install#{@flavour}Debug"
+          DryrunUtils.execute("#{builder} install#{@flavour}Debug")
         end
       end
 
-      if Dryrun::MainApp.getDevice != nil
+      unless @device.nil?
         clear_app_data
         puts "Installing #{@package.green}...\n"
         puts "executing: #{execute_line.green}\n"
@@ -134,10 +134,11 @@ module Dryrun
       end
     end
 
-    def is_gradle_wrapped
-      return false if !File.directory?('gradle/')
+    def gradle_wrapped?
+      return false unless File.directory?('gradle/')
 
-      File.exist?('gradle/wrapper/gradle-wrapper.properties') and File.exist?('gradle/wrapper/gradle-wrapper.jar')
+      File.exist?('gradle/wrapper/gradle-wrapper.properties') &&
+        File.exist?('gradle/wrapper/gradle-wrapper.jar')
     end
 
     def sample_project
@@ -156,69 +157,65 @@ module Dryrun
       [false, false]
     end
 
-    def get_uninstall_command
-     "adb uninstall \"#{@package}\""
-   end
-
-   def clear_app_data
-     DryrunUtils.run_adb("shell pm clear #{@package}")
-   end
-
-   def uninstall_application
-    DryrunUtils.run_adb("shell pm uninstall #{@package}")
-  end
-
-  def get_execution_line_command(path_to_sample)
-    manifest_file = get_manifest(path_to_sample)
-
-    if manifest_file.nil?
-      return false
+    def uninstall_command
+      "adb uninstall \"#{@package}\""
     end
 
-    doc = Oga.parse_xml(manifest_file)
-
-    @package = get_package(doc)
-    @launcher_activity = get_launcher_activity(doc)
-
-    if !@launcher_activity
-      return false
+    def clear_app_data
+      DryrunUtils.run_adb("shell pm clear #{@package}")
     end
 
-    manifest_file.close
+    def uninstall_application
+      DryrunUtils.run_adb("shell pm uninstall #{@package}")
+    end
 
-    return "am start -n \"#{get_launchable_activity}\" -a android.intent.action.MAIN -c android.intent.category.LAUNCHER"
-  end
+    def get_execution_line_command(path_to_sample)
+      manifest_file = get_manifest(path_to_sample)
 
-  def get_manifest(path_to_sample)
-    default_path = File.join(path_to_sample, 'src/main/AndroidManifest.xml')
-    if File.exist?(default_path)
-      return File.open(default_path)
-    else
-      Find.find(path_to_sample) do |path|
-        return File.open(path) if path =~ /.*AndroidManifest.xml$/
+      return false if manifest_file.nil?
+
+      doc = Oga.parse_xml(manifest_file)
+
+      @package = get_package(doc)
+      @launcher_activity = get_launcher_activity(doc)
+
+      return false unless @launcher_activity
+
+      manifest_file.close
+
+      "am start -n \"#{launchable_activity}\" -a android.intent.action.MAIN -c android.intent.category.LAUNCHER"
+    end
+
+    def get_manifest(path_to_sample)
+      default_path = File.join(path_to_sample, 'src/main/AndroidManifest.xml')
+      if File.exist?(default_path)
+        return File.open(default_path)
+      else
+        Find.find(path_to_sample) do |path|
+          return File.open(path) if path =~ /.*AndroidManifest.xml$/
+        end
       end
     end
-  end
 
-  def get_launchable_activity
-    full_path_to_launcher = "#{@package}#{@launcher_activity.gsub(@package,'')}"
-    "#{@package}/#{full_path_to_launcher}"
-  end
+    def launchable_activity
+      full_path_to_launcher = "#{@package}#{@launcher_activity.gsub(@package, '')}"
+      "#{@package}/#{full_path_to_launcher}"
+    end
 
-  def get_package(doc)
-   doc.xpath("//manifest").attr('package').first.value
- end
+    def get_package(doc)
+      doc.xpath('//manifest').attr('package').first.value
+    end
 
- def get_launcher_activity(doc)
-  activities = doc.css('activity')
-  activities.each do |child|
-    intent_filter = child.css('intent-filter')
+    def get_launcher_activity(doc)
+      activities = doc.css('activity')
+      activities.each do |child|
+        intent_filter = child.css('intent-filter')
 
-    if intent_filter != nil and intent_filter.length != 0
-      return child.attr("android:name").value
+        if !intent_filter.nil? && !intent_filter.empty?
+          return child.attr('android:name').value
+        end
+      end
+      false
     end
   end
-  false
-end
-end
 end
